pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-devops-ai-pr-review"
version = "0.1.0"
description = "AI-powered Pull Request review extension for Azure DevOps"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Asif Zaidov", email = "zaidov012@gmail.com"}
]
keywords = ["azure-devops", "ai", "code-review", "pull-request", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers"
]

dependencies = [
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "azure-devops>=7.1.0b4",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "ollama>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "colorlog>=6.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "responses>=0.25.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.0",
]

[project.scripts]
azdo-ai-review = "src.main:main"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
